+++ src/router.ts
@@ -1,31 +1,10 @@
 import { initTRPC } from '@trpc/server'
-import { z } from 'zod'
-import { prisma } from './prisma'
+import { UserRouter } from './routes/User'
 
-const t = initTRPC.create()
+export const t = initTRPC.create()
 
 export const appRouter = t.router({
-  getUser: t.procedure.input(z.string()).query((opts) => {
-    opts.input
-    return { id: opts.input, name: 'Bilbo' }
-  }),
-  createUser: t.procedure
-    .input(
-      z.object({
-        firstName: z.string(),
-        lastName: z.string(),
-        email: z.string(),
-      })
-    )
-    .mutation(async (opts) => {
-      const { input } = opts
-
-      return await prisma.user.create({
-        data: {
-          ...input,
-        },
-      })
-    }),
+  user: UserRouter,
 })
 
 export type AppRouter = typeof appRouter
